Split Settings presenter from oversized files
The CI file-length guard rejected the first cut because the Settings presenter and regression coverage grew already-budgeted files. Moving the presenter and tests into focused small files keeps the ownership fix while reducing existing file length debt.

Constraint: Local app builds and test suites are disallowed for this task; CI owns full validation.\nRejected: Increase the Swift file-length budget | accepts debt when the code can be split cleanly.\nConfidence: high\nScope-risk: narrow\nDirective: Keep Settings presenter lifecycle code out of cmuxApp.swift.\nTested: git diff --check; python3 scripts/swift_file_length_budget.py --budget .github/swift-file-length-budget.tsv\nNot-tested: Local build/test suite per task constraints.
